refactor(storage): move bedrockEmbedder to storage/embeddings/
Separate embedder implementations from search strategies: search/ holds
strategies, embeddings/ holds provider-specific embedder factories.
