+++ strands-ts/src/storage/search/index.ts
@@ -13,5 +13,3 @@ export type { StorageSearchResult } from '../storage.js'
 export { KeywordSearchStrategy } from './keyword.js'
 export { S3VectorSearchStrategy } from './s3-vector.js'
 export type { S3VectorSearchStrategyConfig } from './s3-vector.js'
-export { bedrockEmbedder } from './bedrock-embedder.js'
-export type { BedrockEmbedderConfig } from './bedrock-embedder.js'
